# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /roi/calculate:
    post:
      operationId: calculateRoi
      tags: [roi]
      summary: Compute charter ROI for a yacht
      description: |
        Runs the Charter ROI engine for the user's yacht. Pricing is set by
        `pricing_mode` — manual_daily / manual_weekly require `manual_rate_eur`
        and `manual_charter_units`; ai mode estimates rate via web search and
        falls back to a deterministic regional heuristic if the AI call fails.
        Result is persisted to roi_calculations.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoiCalculationInput"
      responses:
        "200":
          description: ROI computed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoiCalculation"
        "400":
          description: Invalid input (bad shape, manual rate/units out of range)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Yacht not found or not owned by caller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Engine failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: ROI storage not configured (Supabase missing)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /roi/calculations:
    get:
      operationId: listRoiCalculations
      tags: [roi]
      summary: List user's saved ROI calculations
      security:
        - bearerAuth: []
      parameters:
        - name: yacht_id
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: ROI list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoiCalculationListResponse"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /roi/calculations/{id}:
    get:
      operationId: getRoiCalculation
      tags: [roi]
      summary: Get a saved ROI calculation by id
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ROI detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoiCalculationDetail"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CharterRegion:
      type: string
      enum:
        - mediterranean
        - caribbean
        - northern_europe
        - asia_pacific_me
        - middle_east
    CharterSeason:
      type: string
      enum: [high, shoulder, low, mixed]
    ManagementStyle:
      type: string
      enum: [owner_operated, management_company, brokerage]
    OccupancyTarget:
      type: string
      enum: [conservative, realistic, optimistic]
    PricingMode:
      type: string
      enum: [manual_daily, manual_weekly, ai]
      description: |
        How the charter rate is determined:
        * manual_daily  — user provides daily rate + number of charter days
        * manual_weekly — user provides weekly rate + number of charter weeks
        * ai            — engine asks AI (web-search) for plausible rate + weeks
...
    RoiCalculationInput:
      type: object
      required: [yacht_id, region, pricing_mode, management_style]
      properties:
        yacht_id:
          type: string
        region:
          $ref: "#/components/schemas/CharterRegion"
        season:
          oneOf:
            - $ref: "#/components/schemas/CharterSeason"
            - type: "null"
          description: Default 'mixed' (weighted across high/shoulder/low)
        management_style:
          $ref: "#/components/schemas/ManagementStyle"
        occupancy_target:
          oneOf:
            - $ref: "#/components/schemas/OccupancyTarget"
            - type: "null"
          description: Hint for AI mode; ignored in manual modes
        pricing_mode:
          $ref: "#/components/schemas/PricingMode"
        manual_rate_eur:
          type: ["number", "null"]
          minimum: 0
          description: Per-day rate for manual_daily, per-week rate for manual_weekly
        manual_charter_units:
          type: ["integer", "null"]
          minimum: 0
          maximum: 366
          description: Number of charter days OR weeks per year, matching pricing_mode
        management_fee_pct:
          type: ["number", "null"]
          minimum: 0
          maximum: 50
          description: Override default management fee; null = uses style default
        target_weeks:
          type: ["integer", "null"]
          minimum: 0
          maximum: 52
          description: AI mode only — override AI-predicted charter weeks
    RoiComparable:
      type: object
      required: [name]
      properties:
        name:
          type: string
        location:
          type: ["string", "null"]
        weekly_rate_eur:
          type: ["number", "null"]
        source_url:
          type: ["string", "null"]
    ExpenseBreakdown:
      type: object
      required: [category, amount_eur]
      properties:
        category:
          type: string
        amount_eur:
          type: number
        formula:
          type: ["string", "null"]
    MonthlyPoint:
      type: object
      required: [month, value_eur]
      properties:
        month:
          type: integer
          minimum: 1
          maximum: 12
        value_eur:
          type: number
    YearlyPoint:
      type: object
      required: [year_offset, value_eur]
      properties:
        year_offset:
          type: integer
        value_eur:
          type: number
    RoiCalculation:
      type: object
      required:
        - annual_revenue_eur
        - annual_expenses_eur
        - net_profit_eur
        - roi_pct
        - payback_years
        - occupancy_pct
        - expected_charter_weeks
        - avg_daily_rate_eur
        - currency
        - expenses
        - revenue_by_month
        - depreciation_curve
        - roi_projection_5y
        - reasoning
        - confidence
        - legal_disclaimer
      properties:
        id:
          type: ["string", "null"]
        annual_revenue_eur:
          type: number
        annual_expenses_eur:
          type: number
        net_profit_eur:
          type: number
        roi_pct:
          type: number
        payback_years:
          type: number
        occupancy_pct:
          type: number
        expected_charter_weeks:
          type: integer
        avg_daily_rate_eur:
          type: number
        daily_rate_low_season_eur:
          type: ["number", "null"]
        daily_rate_high_season_eur:
          type: ["number", "null"]
        market_rating:
          type: ["string", "null"]
          enum: [A, B, C, D, null]
        risk_score:
          type: ["integer", "null"]
          minimum: 1
          maximum: 10
        currency:
          type: string
        expenses:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseBreakdown"
        revenue_by_month:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyPoint"
        depreciation_curve:
          type: array
          items:
            $ref: "#/components/schemas/YearlyPoint"
        roi_projection_5y:
          type: array
          items:
            $ref: "#/components/schemas/YearlyPoint"
        comparables:
          type: array
          items:
            $ref: "#/components/schemas/RoiComparable"
        reasoning:
          type: string
        recommendations:
          type: array
          items:
            type: string
        confidence:
          type: string
          enum: [high, medium, low]
        legal_disclaimer:
          type: string
    RoiCalculationListItem:
      type: object
      required:
        - id
        - yacht_id
        - created_at
        - region
        - annual_revenue_eur
        - net_profit_eur
        - roi_pct
      properties:
        id:
          type: string
        yacht_id:
          type: string
        created_at:
          type: string
        region:
          type: string
        annual_revenue_eur:
          type: number
        annual_expenses_eur:
          type: ["number", "null"]
        net_profit_eur:
          type: number
        roi_pct:
          type: number
        payback_years:
          type: ["number", "null"]
    RoiCalculationListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RoiCalculationListItem"
    RoiCalculationDetail:
      type: object
      required: [id, yacht_id, created_at, input, result]
      properties:
        id:
          type: string
        yacht_id:
          type: string
        created_at:
          type: string
        input:
          $ref: "#/components/schemas/RoiCalculationInput"
        result:
          $ref: "#/components/schemas/RoiCalculation"