# lib/api-zod/src/generated/types/charterSeason.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type CharterSeason = (typeof CharterSeason)[keyof typeof CharterSeason];

export const CharterSeason = {
  high: "high",
  shoulder: "shoulder",
  low: "low",
  mixed: "mixed",
} as const;