# lib/api-zod/src/generated/types/pricingMode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * How the charter rate is determined:
* manual_daily  — user provides daily rate + number of charter days
* manual_weekly — user provides weekly rate + number of charter weeks
* ai            — engine asks AI (web-search) for plausible rate + weeks

 */
export type PricingMode = (typeof PricingMode)[keyof typeof PricingMode];

export const PricingMode = {
  manual_daily: "manual_daily",
  manual_weekly: "manual_weekly",
  ai: "ai",
} as const;