# lib/api-zod/src/generated/types/roiCalculation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseBreakdown } from "./expenseBreakdown";
import type { MonthlyPoint } from "./monthlyPoint";
import type { RoiCalculationConfidence } from "./roiCalculationConfidence";
import type { RoiCalculationMarketRating } from "./roiCalculationMarketRating";
import type { RoiComparable } from "./roiComparable";
import type { YearlyPoint } from "./yearlyPoint";

export interface RoiCalculation {
  /** @nullable */
  id?: string | null;
  annual_revenue_eur: number;
  annual_expenses_eur: number;
  net_profit_eur: number;
  roi_pct: number;
  payback_years: number;
  occupancy_pct: number;
  expected_charter_weeks: number;
  avg_daily_rate_eur: number;
  /** @nullable */
  daily_rate_low_season_eur?: number | null;
  /** @nullable */
  daily_rate_high_season_eur?: number | null;
  /** @nullable */
  market_rating?: RoiCalculationMarketRating;
  /**
   * @minimum 1
   * @maximum 10
   * @nullable
   */
  risk_score?: number | null;
  currency: string;
  expenses: ExpenseBreakdown[];
  revenue_by_month: MonthlyPoint[];
  depreciation_curve: YearlyPoint[];
  roi_projection_5y: YearlyPoint[];
  comparables?: RoiComparable[];
  reasoning: string;
  recommendations?: string[];
  confidence: RoiCalculationConfidence;
  legal_disclaimer: string;
}