# lib/api-zod/src/generated/types/roiComparable.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RoiComparable {
  name: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  weekly_rate_eur?: number | null;
  /** @nullable */
  source_url?: string | null;
}